Fix un-pageable array results (no href is present), such as invoice/transactions
Certain nested array results are not always part of a pageable collection, so when
we parse them into a Recurly_Pager object, we need to pre-load the _count property
so we can iterate properly.

In order to properly implement this, an accurate count of XML node->childNodes was
required, which was being thrown off because we weren't ignoring whitespace nodes.
This change then required a slight rewrite of the parser.

Acked-by: Stephen Celis <stephen@recurly.com>
Signed-off-by: Evan Owen <kainosnoema@gmail.com>