Transaction refund API route was changed to DELETE /v2/transactions/:uuid
The old route was deprecated some time ago, but somehow this client lib
wasn't updated. I'll be attempting to keep it current from now on.

In addition to changing the refund route, we're now parsing any responses
that might be returned by DELETE requests, as is the case with refunding
transactions.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>