Fix updating currency lists on loaded resources, such as plans
Fixes an issue where updating a specific currency on a previously
loaded resource wouldn't take affect. Because the currency list
is neither an array nor updates the unsavedAttrs array on the parent,
it wasn't being persisted. This checks if the attr in question is
either an array or implements ArrayAccess, in which case it's
persisted automatically.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>