bulk param
This new parameter allows you to bypass the 60 second limit on adding
multiple subscriptions for the same plan to one account.

The boolean serialization to {0,1} was changed to {true,false} to match
the api. It was not being used by any of the other serializers so should
be backwards compatible.

Also adding a test for $account->tax_exempt to make sure false is
working correctly.