Adding vat_location_valid field to Account
This field will show up when the account is elligible and has been
location verified according to the VAT rules. It will be true when
passed and false when failed.

There weren't any needed changes for the library code here as it's a
read-only attribute. I just wanted to update the fixture and the test.