Return correct node name for Recurly_NoteList
Notes weren't loaded after rewind() was called. The name didn't match what
was in the XML so __parseXmlToObject() was never called and hence the object
were never added to $this->_objects.

While I'm here I fixed a bug in the pager where it'd keep checking for a
next link so it could reload... that wasn't ever going to happen so it's
better to give a hard error.