Drop older versions of PHP and HHVM
Per the README file, this package only supports versions of PHP that are currently supported by the project. As of a few days ago, that's PHP 7.2 and above.

There was also mention of HHVM, but only the LTS releases that still support PHP; unfortunately, that appears to have ended some time ago.